#000 fix: CSP Refactor
